Change H3Index interop type to c_ulonglong
c_long will be the long type, which on Windows is 32 bits. This resulted
in some methods returning broken data (half of the H3Index bits) and in
other methods segfaulting (the polyfill call in the compact test caused
all tests to exit immediately).

Minor adjustment to build script to use `cmake --build` instead of
calling `msbuild.exe` directly.

Adjusted a test to not assert on floating point equality.